requirements.txt
## Requests
feedparser==5.2.1
requests==2.9.1

## Database
pymongo==3.2.1
mongoengine==0.10.6
blinker==1.4

## Parsing
beautifulsoup4==4.4.1
lxml==3.5.0

## Configuration
confire==0.2.0
PyYAML==3.11

## Command Line
commis==0.2
colorama==0.3.6

## Utilities
schedule==0.3.2
python-dateutil==2.4.2
six==1.10.0

## Testing
## Uncomment and install for development
#nose==1.3.7
#coverage==4.0.3
#mock==1.3.0
#funcsigs==0.4
#pbr==1.8.1
#mongomock==3.2.1
#sentinels==0.0.6

## Building
## Uncomment and install for deployment
#wheel==0.29.0
